OAK-10114: Node diff with many changes uses excessive memory
Move tests to MongoDocumentStoreIT
